+++ History.md
@@ -4,6 +4,7 @@
   * updated; hooks.js to 0.2.1
   * fixed; issue with passing undefined to a hook callback. thanks to [chrisleishman] for reporting.
   * fixed; updating/setting nested objects in strict schemas (#843) as reported by [kof]
+  * fixed; Query#{update,remove}() work without callbacks again (#788)
 
 2.5.14 / 2012-04-13
 ===================